Use common StateTable for all dynamic algorithms
