+++ rustfst/src/algorithms/cache/mod.rs
@@ -1,9 +1,11 @@
 pub use self::cache_impl::CacheImpl;
 pub use self::cache_state::CacheState;
 pub use self::fst_impl::FstImpl;
+pub use self::state_table::StateTable;
 pub use self::vector_cache_state::VectorCacheState;
 
 mod cache_impl;
 mod cache_state;
 mod fst_impl;
+mod state_table;
 mod vector_cache_state;
